+++ src/llm_orc/cli.py
@@ -2,6 +2,12 @@
 
 import click
 
+from llm_orc.cli_commands import (
+    invoke_ensemble,
+    list_ensembles_command,
+    list_profiles_command,
+    serve_ensemble,
+)
 from llm_orc.cli_modules.commands.auth_commands import (
     add_auth_provider,
     list_auth_providers,
@@ -13,12 +19,8 @@ from llm_orc.cli_modules.commands.config_commands import (
     check_global_config,
     check_local_config,
     init_local_config,
-    invoke_ensemble,
-    list_ensembles_command,
-    list_profiles_command,
     reset_global_config,
     reset_local_config,
-    serve_ensemble,
 )
