+++ src/llm_orc/integrations/mcp/server.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from llm_orc.core.config.config_manager import ConfigurationManager
 from llm_orc.core.config.ensemble_config import EnsembleConfig, EnsembleLoader
-from llm_orc.ensemble_execution import EnsembleExecutor
+from llm_orc.core.execution.ensemble_execution import EnsembleExecutor
 
 
 class MCPServer:
